Moving feed data from Feed component to its own store
All logic dealing with the updating of data in the list view
has been moved out to its own store.  Still need to update the
Map component to make use of this new store.  This will update
should dramatically reduce the amount of data moving over the
wire as it only needs to communicate the *new* checkins.